Ph1 quarter final

+++ Restaurant/Rest/Order.h
@@ -23,7 +23,7 @@ protected:
 
 	
 	/*Start sir_sayed modification*/
-		int Priorty;
+		double  Priorty;
 	/*End sir_sayed modification*/
 
 public:
